Reorganize tests by benchmark ownership
Group runner and metric tests while consolidating shared setup to keep the suite line-negative and easier to navigate.
