+++ contributing.md
@@ -306,7 +306,7 @@ Once your submission PR looks ready, a maintainer applies the `ingest-model` lab
 
 The secret-bearing ingestion workflow never executes PR code. PRs that change the payload *format* (new keys/metrics or schema tweaks in generator code) must regenerate payloads locally and commit them in the same PR:
 
-- Roster drift fails `tests/test_fig_payloads.py`'s coverage tests (e.g. `test_discovery_payload_covers_active_models` after adding or superseding a model). Fix by running `uv run --with-editable . scripts/ingest_model.py <model_name> --payloads-only` and committing the changed `site/src/figs/*.jsonl` files — this splices only your model's freshly computed entries into the committed payloads, pulling your prediction files from the `pred_file_url` entries in the model YAML (the same links automated ingestion uses).
+- Roster drift fails `tests/site/test_fig_payloads.py`'s coverage tests (e.g. `test_discovery_payload_covers_active_models` after adding or superseding a model). Fix by running `uv run --with-editable . scripts/ingest_model.py <model_name> --payloads-only` and committing the changed `site/src/figs/*.jsonl` files — this splices only your model's freshly computed entries into the committed payloads, pulling your prediction files from the `pred_file_url` entries in the model YAML (the same links automated ingestion uses).
 - The `model-pr-guard` check fails any PR that changes payload-generating code without also updating committed payloads (`site/src/figs` data files or the route-local `per-element-each-errors.jsonl`): regenerate locally with `scripts/ingest_model.py --payloads-only` (no model name needed for pure format changes), or add the `payloads-unchanged` label for output-neutral refactors.
 - Model-independent data-page payloads and route-local element counts are owned by `scripts/export_data_fig_payloads.py`. Run it after changing that script or `matbench_discovery/data_figs.py`; it requires the local `data/mp/2022-09-16-mp-trj-summary.json.bz2` cache. If absent, regenerate that cache from the extXYZ source with `python data/mp/eda_mp_trj.py`.
