+++ tests/scripts/test_build_diatomic_reference.py
