+++ tests/site/test_parity_assets.py
